resutls.html: Display API request link
As making the talk to QECamp, we realized with @ginger
this might be handy actually.

For now extract UUID from the baseUrl, but it is marely a hack.

We should think about how to expose this metadata for each run,
so it will work also if we have multiple API endpoints
(e.g dev, staging, ...). Felt parsing this out of `pipeline.log`
or some other places it is even worse.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>