results.html: Fix wrong inline logs when showing successful tests
It's not enough to load LogViewer's logs for the first instantiation
only. When showing successful logs, and there is a successful test plan
*before* the first failed one, then apparently browsers keep the first
original existing `<log-viewer>` (previously showing the first failed
plan) instance at the same place in the DOM, and update its properties
to show the successful plan. Thus, previously it would still show the
contents of the first failed plan.

To fix this, reload the log every time the `url` attribute changes, not
just the first time. This is also conceptually cleaner.

Add the advanced plan reproducer to our tmt-mixed scenario, and check
the expected tmt-reproducer link and inline contents.

Closes #3