Show HTML artifacts inline as iframes
Tests can now attach custom artifacts via tmt 1.18's new "custom
results.yaml" feature [1]. Common cases are a specialized viewer for
this kind of test, or things like a code coverage report. These are
useful to display inline, so render them as iframe.

If there are HTML artifacts, don't render any other test logs inline, assuming
that the more specific viewer will take care of that. That reduces
distraction/clutter and increases screen space.

The iframe height is a bit difficult: When it is on a different origin,
we cannot even inspect it; and even if we can, the size changes
dynamically, and it's hard to follow events from an iframe. So just pin
it at 80% of our own page height, which is a good enough initial viewer.
For more viewing space, users can always click the link to get it on a
full page.

[1] https://tmt.readthedocs.io/en/latest/spec/tests.html#result