Move logs down after tests and inline artifacts
Put more emphasis on the test specific outputs by moving them up on the
page.

Move inline artifacts on top of log links. For tests this is primarily
the "testout.log", which is the most interesting thing on the test
level.

Move both inline artifacts like tmt-reproducer and further log links to
the bottom of a plan, below the test list. To keep them discoverable,
add a "Go to Logs and Artifacts" link to the top of a plan, as when
there are many tests, these can now be quite far away from the top of
the plan element.