test: Stop assuming log-viewer ordering
`cy.get()` returns the first matching element, thus the test was
previously assuming a particular ordering of `<log-viewer>` components.
We are going to change that.

Select the desired log-viewer more carefully by the URL that it shows.
Some tests were already doing that, and it was already done correctly
for the second, third, etc. log-viewer.